+++ projects/rocprofiler-compute/src/rocprof_compute_analyze/analysis_webui.py
@@ -247,7 +247,10 @@ class webui_analysis(OmniAnalyze_Base):
             # Generate GUI content
             # ~~~~~~~~~~~~~~~~~~~~~~~
             div_children = [
-                get_memchart(panel_configs[300]["data source"], base_data[base_run])
+                get_memchart(
+                    panel_configs.get(300, {}).get("data source"),
+                    base_data[base_run],
+                )
             ]
 
             is_roofline_valid, roofline_error_msg = validate_roofline_csv(